feat(eth): Move event renderers to their own package
