+++ eth/events/output_full_test.go
@@ -1,4 +1,4 @@
-package eth
+package events
 
 import (
 	"encoding/hex"
